chore: merge main and fix resulting problems
